Bug Fixing For selecting from produit
i cahnged the * to specifed culomns

+++ UBonRecGestion.pas
@@ -406,17 +406,17 @@ begin
 
        if ResherchPARDesProduitsRdioBtn.Checked then
    begin
-        MainForm.ProduitTable.Active:=False;
-        MainForm.ProduitTable.SQL.Clear;
-        MainForm.ProduitTable.SQL.Text:= 'SELECT * FROM produit WHERE LOWER(nom_p) LIKE LOWER('+QuotedStr(ProduitBonRecGCbx.Text)+')';
-        MainForm.ProduitTable.Active:=True;
-        CodeP:= MainForm.ProduitTable.FieldByName('code_p').AsInteger ;
+        MainForm.SQLQuery.Active:=False;
+        MainForm.SQLQuery.SQL.Clear;
+        MainForm.SQLQuery.SQL.Text:= 'SELECT code_p,nom_p,prixht_p,prixvd_p,prixvr_p,prixvg_p,prixva_p,prixva2_p,tva_p,perissable_p FROM produit WHERE LOWER(nom_p) LIKE LOWER('+QuotedStr(ProduitBonRecGCbx.Text)+')';
+        MainForm.SQLQuery.Active:=True;
+        CodeP:= MainForm.SQLQuery.FieldByName('code_p').AsInteger ;
 
          lookupResultRefP := MainForm.Bona_recPlistTable.Lookup('code_p',(CodeP),'code_p');
          if VarIsnull( lookupResultRefP) then
          begin
 
-            if  MainForm.ProduitTable.RecordCount > 0  then
+            if  MainForm.SQLQuery.RecordCount > 0  then
           begin
 
             MainForm.Bona_recPlistTable.DisableControls;
@@ -437,7 +437,7 @@ begin
                   CodeBR:= MainForm.Bona_recPlistTable.FieldValues['code_barecl'] + 1 ;
                  end;
 
-                 if MainForm.ProduitTable.FieldByName('perissable_p').AsBoolean = True then
+                 if MainForm.SQLQuery.FieldByName('perissable_p').AsBoolean = True then
                  begin
                    ProduitsListDBGridEh.Columns[4].Visible := True
                  end;
@@ -446,16 +446,16 @@ begin
              MainForm.Bona_recPlistTable.Append;
              MainForm.Bona_recPlistTable.FieldValues['code_barecl']:= CodeBR ;
              MainForm.Bona_recPlistTable.FieldValues['code_barec']:= MainForm.Bona_recTable.FieldValues['code_barec'];
-             MainForm.Bona_recPlistTable.FieldValues['code_p']:=  MainForm.ProduitTable.FieldValues['code_p'] ;
+             MainForm.Bona_recPlistTable.FieldValues['code_p']:=  MainForm.SQLQuery.FieldValues['code_p'] ;
              MainForm.Bona_recPlistTable.FieldValues['qut_p'] :=  01;
-             MainForm.Bona_recPlistTable.FieldValues['prixht_p']:= MainForm.ProduitTable.FieldValues['prixht_p'];
+             MainForm.Bona_recPlistTable.FieldValues['prixht_p']:= MainForm.SQLQuery.FieldValues['prixht_p'];
              MainForm.Bona_recPlistTable.FieldValues['cond_p']:= 01;
-             MainForm.Bona_recPlistTable.FieldValues['tva_p']:= MainForm.ProduitTable.FieldValues['tva_p'];
-             MainForm.Bona_recPlistTable.FieldByName('prixvd_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvd_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixvr_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvr_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixvg_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvg_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixva_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixva_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixva2_p').AsCurrency:= MainForm.ProduitTable.FieldByName('prixva2_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldValues['tva_p']:= MainForm.SQLQuery.FieldValues['tva_p'];
+             MainForm.Bona_recPlistTable.FieldByName('prixvd_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvd_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixvr_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvr_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixvg_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvg_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixva_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixva_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixva2_p').AsCurrency:= MainForm.SQLQuery.FieldByName('prixva2_p').AsCurrency;
 
              MainForm.Bona_recPlistTable.FieldValues['qutinstock_p']:= 
              (MainForm.Bona_recPlistTable.FieldValues['qut_p'])*(MainForm.Bona_recPlistTable.FieldValues['cond_p']);     
@@ -502,7 +502,7 @@ begin
       FSplashAddUnite.CancelAddUniteSBtn.Left:= ((FSplashAddUnite.Width div 2 )+((FSplashAddUnite.Width div 2)div 2 ) ) - (FSplashAddUnite.CancelAddUniteSBtn.Width div 2) - 15;
       if  MainForm.Bona_recPlistTable.FieldValues['code_p'] <> NULL then
       begin
-      NomP:=   MainForm.ProduitTable.FieldValues['nom_p'];
+      NomP:=   MainForm.SQLQuery.FieldValues['nom_p'];
       end else begin
         NomP:='';
       end;
@@ -534,16 +534,16 @@ begin
 //--------------------------------------------------------------------------------------------------------------------
    if ResherchPARRefProduitsRdioBtn.Checked then
   begin
-            MainForm.ProduitTable.Active:=False;
-            MainForm.ProduitTable.SQL.Clear;
-            MainForm.ProduitTable.SQL.Text:= 'SELECT * FROM produit WHERE LOWER(refer_p) LIKE LOWER('+QuotedStr(ProduitBonRecGCbx.Text)+')';
-            MainForm.ProduitTable.Active:=True;
-            CodeP:= MainForm.ProduitTable.FieldByName('code_p').AsInteger ;
+            MainForm.SQLQuery.Active:=False;
+            MainForm.SQLQuery.SQL.Clear;
+            MainForm.SQLQuery.SQL.Text:= 'SELECT code_p,refer_p,nom_p,prixht_p,prixvd_p,prixvr_p,prixvg_p,prixva_p,prixva2_p,tva_p,perissable_p FROM produit WHERE LOWER(refer_p) LIKE LOWER('+QuotedStr(ProduitBonRecGCbx.Text)+')';
+            MainForm.SQLQuery.Active:=True;
+            CodeP:= MainForm.SQLQuery.FieldByName('code_p').AsInteger ;
 
          lookupResultRefP := MainForm.Bona_recPlistTable.Lookup('code_p',(CodeP),'code_p');
          if VarIsnull( lookupResultRefP) then
          begin
-       if  MainForm.ProduitTable.RecordCount > 0  then
+       if  MainForm.SQLQuery.RecordCount > 0  then
           begin
 
             MainForm.Bona_recPlistTable.DisableControls;
@@ -561,7 +561,7 @@ begin
                 CodeBR:= MainForm.Bona_recPlistTable.FieldValues['code_barecl'] + 1 ;
                end;
 
-                 if MainForm.ProduitTable.FieldByName('perissable_p').AsBoolean = True then
+                 if MainForm.SQLQuery.FieldByName('perissable_p').AsBoolean = True then
                  begin
                    ProduitsListDBGridEh.Columns[4].Visible := True
                  end;
@@ -569,20 +569,20 @@ begin
              MainForm.Bona_recPlistTable.Insert;
              MainForm.Bona_recPlistTable.FieldValues['code_barecl']:= CodeBR ;
              MainForm.Bona_recPlistTable.FieldValues['code_barec']:= MainForm.Bona_recTable.FieldValues['code_barec'];
-             MainForm.Bona_recPlistTable.FieldValues['code_p']:=  MainForm.ProduitTable.FieldValues['code_p'] ;
+             MainForm.Bona_recPlistTable.FieldValues['code_p']:=  MainForm.SQLQuery.FieldValues['code_p'] ;
              MainForm.Bona_recPlistTable.FieldValues['qut_p'] :=  01;
-             MainForm.Bona_recPlistTable.FieldValues['prixht_p']:= MainForm.ProduitTable.FieldValues['prixht_p'];
+             MainForm.Bona_recPlistTable.FieldValues['prixht_p']:= MainForm.SQLQuery.FieldValues['prixht_p'];
              MainForm.Bona_recPlistTable.FieldValues['cond_p']:= 01;
-             MainForm.Bona_recPlistTable.FieldValues['tva_p']:= MainForm.ProduitTable.FieldValues['tva_p'];
-             MainForm.Bona_recPlistTable.FieldByName('prixvd_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvd_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixvr_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvr_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixvg_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvg_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixva_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixva_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixva2_p').AsCurrency:= MainForm.ProduitTable.FieldByName('prixva2_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldValues['tva_p']:= MainForm.SQLQuery.FieldValues['tva_p'];
+             MainForm.Bona_recPlistTable.FieldByName('prixvd_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvd_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixvr_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvr_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixvg_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvg_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixva_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixva_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixva2_p').AsCurrency:= MainForm.SQLQuery.FieldByName('prixva2_p').AsCurrency;
 
               MainForm.Bona_recPlistTable.FieldValues['qutinstock_p']:= 
              (MainForm.Bona_recPlistTable.FieldValues['qut_p'])*(MainForm.Bona_recPlistTable.FieldValues['cond_p']);     
-             
+
              MainForm.Bona_recPlistTable.Post ;
              MainForm.Bona_recPlistTable.IndexFieldNames:='code_barec';
 
@@ -624,7 +624,7 @@ begin
             FSplashAddUnite.CancelAddUniteSBtn.Left:= ((FSplashAddUnite.Width div 2 )+((FSplashAddUnite.Width div 2)div 2 ) ) - (FSplashAddUnite.CancelAddUniteSBtn.Width div 2) - 15;
             if  MainForm.Bona_recPlistTable.FieldValues['code_p'] <> NULL then
             begin
-            NomP:=   MainForm.ProduitTable.FieldValues['nom_p'];
+            NomP:=   MainForm.SQLQuery.FieldValues['nom_p'];
             end else begin
               NomP:='';
             end;
@@ -663,13 +663,13 @@ begin
     CodeCB:=MainForm.SQLQuery.FieldValues['code_p'];
    end;
 
-    MainForm.ProduitTable.Active:=False;
-    MainForm.ProduitTable.SQL.Clear;
-    MainForm.ProduitTable.SQL.Text:= 'SELECT * FROM produit WHERE code_p = '+QuotedStr(IntToStr(CodeCB)) +'OR'+ ' LOWER(codebar_p) LIKE LOWER(' + QuotedStr(ProduitBonRecGCbx.Text)+')';
-    MainForm.ProduitTable.Active:=True;
-    CodeP:= MainForm.ProduitTable.FieldByName('code_p').AsInteger ;
+    MainForm.SQLQuery.Active:=False;
+    MainForm.SQLQuery.SQL.Clear;
+    MainForm.SQLQuery.SQL.Text:= 'SELECT code_p,nom_p,codebar_p,prixht_p,prixvd_p,prixvr_p,prixvg_p,prixva_p,prixva2_p,tva_p,perissable_p FROM produit WHERE code_p = '+QuotedStr(IntToStr(CodeCB)) +'OR'+ ' LOWER(codebar_p) LIKE LOWER(' + QuotedStr(ProduitBonRecGCbx.Text)+')';
+    MainForm.SQLQuery.Active:=True;
+    CodeP:= MainForm.SQLQuery.FieldByName('code_p').AsInteger ;
 
-     if  (MainForm.ProduitTable.RecordCount > 0 )   then
+     if  (MainForm.SQLQuery.RecordCount > 0 )   then
       begin
 
          lookupResultRefP := MainForm.Bona_recPlistTable.Lookup('code_p',(CodeP),'code_p');
@@ -692,7 +692,7 @@ begin
                 CodeBR:= MainForm.Bona_recPlistTable.FieldValues['code_barecl'] + 1 ;
                end;
                
-                 if MainForm.ProduitTable.FieldByName('perissable_p').AsBoolean = True then
+                 if MainForm.SQLQuery.FieldByName('perissable_p').AsBoolean = True then
                  begin
                    ProduitsListDBGridEh.Columns[4].Visible := True
                  end;               
@@ -700,16 +700,16 @@ begin
              MainForm.Bona_recPlistTable.Insert;
              MainForm.Bona_recPlistTable.FieldValues['code_barecl']:= CodeBR ;
              MainForm.Bona_recPlistTable.FieldValues['code_barec']:= MainForm.Bona_recTable.FieldValues['code_barec'];
-             MainForm.Bona_recPlistTable.FieldValues['code_p']:=  MainForm.ProduitTable.FieldValues['code_p'] ;
+             MainForm.Bona_recPlistTable.FieldValues['code_p']:=  MainForm.SQLQuery.FieldValues['code_p'] ;
              MainForm.Bona_recPlistTable.FieldValues['qut_p'] :=  01;
-             MainForm.Bona_recPlistTable.FieldValues['prixht_p']:= MainForm.ProduitTable.FieldValues['prixht_p'];
+             MainForm.Bona_recPlistTable.FieldValues['prixht_p']:= MainForm.SQLQuery.FieldValues['prixht_p'];
              MainForm.Bona_recPlistTable.FieldValues['cond_p']:= 01;
-             MainForm.Bona_recPlistTable.FieldValues['tva_p']:= MainForm.ProduitTable.FieldValues['tva_p'];
-             MainForm.Bona_recPlistTable.FieldByName('prixvd_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvd_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixvr_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvr_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixvg_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixvg_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixva_p').AsCurrency:=  MainForm.ProduitTable.FieldByName('prixva_p').AsCurrency;
-             MainForm.Bona_recPlistTable.FieldByName('prixva2_p').AsCurrency:= MainForm.ProduitTable.FieldByName('prixva2_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldValues['tva_p']:= MainForm.SQLQuery.FieldValues['tva_p'];
+             MainForm.Bona_recPlistTable.FieldByName('prixvd_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvd_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixvr_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvr_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixvg_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixvg_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixva_p').AsCurrency:=  MainForm.SQLQuery.FieldByName('prixva_p').AsCurrency;
+             MainForm.Bona_recPlistTable.FieldByName('prixva2_p').AsCurrency:= MainForm.SQLQuery.FieldByName('prixva2_p').AsCurrency;
 
              MainForm.Bona_recPlistTable.FieldValues['qutinstock_p']:= 
              (MainForm.Bona_recPlistTable.FieldValues['qut_p'])*(MainForm.Bona_recPlistTable.FieldValues['cond_p']);             
@@ -756,7 +756,7 @@ begin
             FSplashAddUnite.CancelAddUniteSBtn.Left:= ((FSplashAddUnite.Width div 2 )+((FSplashAddUnite.Width div 2)div 2 ) ) - (FSplashAddUnite.CancelAddUniteSBtn.Width div 2) - 15;
             if  MainForm.Bona_recPlistTable.FieldValues['code_p'] <> NULL then
             begin
-            NomP:=   MainForm.ProduitTable.FieldValues['nom_p'];
+            NomP:=   MainForm.SQLQuery.FieldValues['nom_p'];
             end else begin
               NomP:='';
             end;
@@ -785,10 +785,8 @@ begin
           end;
 
    end;
-          MainForm.ProduitTable.Active:=False;
-          MainForm.ProduitTable.SQL.Clear;
-          MainForm.ProduitTable.SQL.Text:= 'SELECT * FROM produit';
-          MainForm.ProduitTable.Active := True;
+          MainForm.SQLQuery.Active:=False;
+          MainForm.SQLQuery.SQL.Clear;
 
          MainForm.Bona_recPlistTable.Refresh;
 //        ProduitBonRecGCbx.AutoDropDown:=False;
@@ -889,6 +887,8 @@ begin
  AND (MainForm.Bona_recTable.FieldByName('code_f').AsInteger <> 0)  then
  begin
    FournisseurBonRecGCbx.Text:= MainForm.Bona_recTable.FieldValues['fourbarec'];
+   ModePaieBonRecGCbx.Text:= MainForm.Bona_recTable.FieldValues['ModePaie'];
+   CompteBonRecGCbx.Text:= MainForm.Bona_recTable.FieldValues['Compte'];
    ProduitBonRecGCbx.SetFocus;
  end else
      begin
@@ -1280,10 +1280,13 @@ procedure TBonRecGestionF.FormCloseQuery(Sender: TObject;
           MainForm.CompteTable.EnableControls;
 
         //------- This is to delete data from tre and reg ih not valide----------------------------------------------
+           if (codeBR <> 0) AND (codeBR <> null) then
+           begin
               MainForm.GstockdcConnection.ExecSQL('DELETE FROM regfournisseur where code_barec = ' + IntToStr(codeBR));
               MainForm.GstockdcConnection.ExecSQL('DELETE FROM opt_cas_bnk where code_barec = ' + IntToStr(codeBR));
               MainForm.RegfournisseurTable.Refresh ;
               MainForm.Opt_cas_bnk_CaisseTable.Refresh ;
+           end;
 
           end;
 
@@ -1412,17 +1415,17 @@ begin
    if MainForm.Bona_recTable.FieldValues['valider_barec'] <> True then
    begin
 
-    MainForm.ProduitTable.DisableControls;
-    MainForm.ProduitTable.Active:=False;
-    MainForm.ProduitTable.SQL.Clear;
-    MainForm.ProduitTable.SQL.Text:='SELECT * FROM produit WHERE code_p = ' +IntToStr(MainForm.Bona_recPlistTable.FieldValues['code_p']);
-    MainForm.ProduitTable.Active:=True;
+//    MainForm.SQLQuery.DisableControls;
+    MainForm.SQLQuery.Active:=False;
+    MainForm.SQLQuery.SQL.Clear;
+    MainForm.SQLQuery.SQL.Text:='SELECT code_p,qut_p,qutini_p FROM produit WHERE code_p = ' +IntToStr(MainForm.Bona_recPlistTable.FieldValues['code_p']);
+    MainForm.SQLQuery.Active:=True;
 
 
     BonRecGOLDStock.Caption:=
-     floatTostrF((MainForm.ProduitTable.FieldValues['QutDispo']),ffNumber,14,2);
+     floatTostrF(((MainForm.SQLQuery.FieldValues['qut_p'] + MainForm.SQLQuery.FieldValues['qutini_p'])),ffNumber,14,2);
     BonRecGNEWStock.Caption:=
-     floatTostrF(((MainForm.ProduitTable.FieldValues['QutDispo'])+((MainForm.Bona_recPlistTable.FieldValues['qut_p']) * (MainForm.Bona_recPlistTable.FieldValues['cond_p']))),ffNumber,14,2);
+     floatTostrF((((MainForm.SQLQuery.FieldValues['qut_p'] + MainForm.SQLQuery.FieldValues['qutini_p']))+((MainForm.Bona_recPlistTable.FieldValues['qut_p']) * (MainForm.Bona_recPlistTable.FieldValues['cond_p']))),ffNumber,14,2);
 
 //     if(StrToFloat (StringReplace(BonRecGNEWStock.Caption, #32, '', [rfReplaceAll])))  < 0 then
 //    begin
@@ -1433,11 +1436,11 @@ begin
 //        Label20.Visible:=false;
 //        end;
 
-    MainForm.ProduitTable.Active:=False;
-    MainForm.ProduitTable.SQL.Clear;
-    MainForm.ProduitTable.SQL.Text:='SELECT * FROM produit ';
-    MainForm.ProduitTable.Active:=True;
-    MainForm.ProduitTable.EnableControls;
+    MainForm.SQLQuery.Active:=False;
+    MainForm.SQLQuery.SQL.Clear;
+//    MainForm.SQLQuery.SQL.Text:='SELECT * FROM produit ';
+//    MainForm.SQLQuery.Active:=True;
+//    MainForm.SQLQuery.EnableControls;
 
 
     RemisePerctageBonRecGEdt.Enabled:=True;
